+++ src/main/java/de/thb/ics/model/Station.java
@@ -10,17 +10,17 @@ import java.time.LocalDate;
 @Builder
 public class Station {
     @Schema(example = "NV141")
-    long id;
+    int id;
 
     @Schema(example = "2006-05-26")
     LocalDate date;
 
     @Schema(example = "42")
-    int target;
+    Integer target;
 
     @Schema(example = "33")
-    int actual;
+    Integer actual;
 
     @Schema(example = "-9")
-    int variance;
+    Integer variance;
 }
